+++ ichnaea/webapp/config.py
@@ -18,17 +18,14 @@ from ichnaea.log import (
 from ichnaea.service import configure_service
 
 
-def main(global_config, app_config=None, init=False,
+def main(global_config, app_config, init=False,
          _db_rw=None, _db_ro=None, _geoip_db=None,
          _raven_client=None, _redis_client=None, _stats_client=None):
 
     configure_logging()
 
-    if app_config is not None:
-        app_settings = app_config.get_map('ichnaea')
-    else:
-        app_settings = {}
-    config = Configurator(settings=app_settings)
+    # make config file settings available
+    config = Configurator(settings=app_config.asdict())
 
     # add support for pt templates
     config.include('pyramid_chameleon')
@@ -39,20 +36,22 @@ def main(global_config, app_config=None, init=False,
     # configure outside connections
     registry = config.registry
 
-    registry.db_rw = configure_db(app_settings.get('db_master'), _db=_db_rw)
-    registry.db_ro = configure_db(app_settings.get('db_slave'), _db=_db_ro)
+    registry.db_rw = configure_db(
+        app_config.get('ichnaea', 'db_master'), _db=_db_rw)
+    registry.db_ro = configure_db(
+        app_config.get('ichnaea', 'db_slave'), _db=_db_ro)
 
     registry.raven_client = raven_client = configure_raven(
-        app_settings.get('sentry_dsn'), _client=_raven_client)
+        app_config.get('ichnaea', 'sentry_dsn'), _client=_raven_client)
 
     registry.redis_client = configure_redis(
-        app_settings.get('redis_url'), _client=_redis_client)
+        app_config.get('ichnaea', 'redis_url'), _client=_redis_client)
 
     registry.stats_client = configure_stats(
-        app_settings.get('statsd_host'), _client=_stats_client)
+        app_config.get('ichnaea', 'statsd_host'), _client=_stats_client)
 
     registry.geoip_db = configure_geoip(
-        app_settings.get('geoip_db_path'), raven_client=raven_client,
+        app_config.get('ichnaea', 'geoip_db_path'), raven_client=raven_client,
         _client=_geoip_db)
 
     config.add_tween('ichnaea.db.db_tween_factory', under=EXCVIEW)